# Use ciso8601 library to parse datetime faster