--- a/homeassistant/package_constraints.txt
+++ b/homeassistant/package_constraints.txt
@@ -7,6 +7,7 @@ async_timeout==3.0.1
 attrs==19.3.0
 bcrypt==3.1.7
 certifi>=2019.11.28
+ciso8601==2.1.3
 cryptography==2.8
 defusedxml==0.6.0
 distro==1.4.0