--- a/pylintrc
+++ b/pylintrc
@@ -5,6 +5,7 @@ ignore=tests
 jobs=2
 load-plugins=pylint_strict_informational
 persistent=no
+extension-pkg-whitelist=ciso8601
 
 [BASIC]
 good-names=id,i,j,k,ex,Run,_,fp