@balloob 2020-02-24
Oh that's a great speedup! 

What do you use to generate the call tree?