@balloob 2020-02-24
I am 👍 on this PR. 

@pvizeli can you comment on if this pacakge can build on our Alpine images.