@pvizeli 2020-02-24
That is just a simple C extension that handles automatically by our wheel infrastructure after this PR is merged.