@KapJI 2020-02-24
To generate calltree I use cProfile and wrap `humanify` call with:
```
pr = cProfile.Profile()
pr.enable()
result = list(humanify(hass, yield_events(query)))
pr.disable()
pr.create_stats()
pr.dump_stats("logbook.cprof")
return result
```
Then I need to convert calltree to different format using `pyprof2calltree`:
```
pyprof2calltree -i logbook.cprof -o ~/logbook.calltree
```
And it can be viewed with `qcachegrind`:
```
qcachegrind logbook.calltree
```
<img width="1680" alt="Screen Shot 2020-02-24 at 8 24 15 AM" src="https://user-images.githubusercontent.com/3767331/75137263-16e4a380-56df-11ea-9d15-6ea036739d18.png">