@KapJI 2020-02-24
Fix lint warning and test. `ciso8601` accepts just date as a valid string, I hope it's not a big deal.