@balloob 2020-02-27
Nope, that file is maintained. Re-run `pip3 install -e .` to pick it up.