@OnFreund 2020-02-27
I already installed with `requirements_all` so everything is working now, but initially after this PR, following the docs for installing testing requirements did not install ciso8601. If no one else encountered this it could be just a local problem.