@balloob 2020-02-27
that's because as part of the tests we run `pip3 install -e .`, which installs Home Assistant and all of its requirements. They are not part of the test requirements.